delete repository inline

+++ app/components/course-page/delete-repository-modal.ts
@@ -2,9 +2,9 @@ import Component from '@glimmer/component';
 import RepositoryModel from 'codecrafters-frontend/models/repository';
 import RouterService from '@ember/routing/router-service';
 import Store from '@ember-data/store';
-import window from 'ember-window-mock';
 import { action } from '@ember/object';
 import { inject as service } from '@ember/service';
+import { tracked } from '@glimmer/tracking';
 
 interface Signature {
   Element: HTMLDivElement;
@@ -16,6 +16,8 @@ interface Signature {
 }
 
 export default class DeleteRepositoryModalComponent extends Component<Signature> {
+  @tracked isDeleting: boolean = false;
+
   @service declare router: RouterService;
   @service declare store: Store;
 
@@ -25,9 +27,15 @@ export default class DeleteRepositoryModalComponent extends Component<Signature>
       return;
     }
 
-    this.router.transitionTo('catalog');
-    await this.args.repository.destroyRecord();
-    window.location.reload();
+    if (this.isDeleting) {
+      return;
+    }
+
+    this.isDeleting = true;
+
+    const trackSlug = this.args.repository.language?.slug; // Store this before we destroy the record
+    await this.args.repository.destroyRecord(); // TODO: Add failure handling
+    this.router.transitionTo('course.introduction', { queryParams: { repo: 'new', track: trackSlug } }).followRedirects();
   }
 }
